Don't fail if the server sends an unknown field
Currently, if the API is changed, and something new is added, this causes a failure, which currently requires a new release to fix. This allows it to work silently, so that it will work, just won't have the new data. This keeps an addition to their API from taking down everyone using the gem.